--- ProjectPlanner/events.html
@@ -1,44 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Events!</title>
-    <link rel="stylesheet" href="assets/styles/events.css" />
-    <script src="assets/scripts/events.js" defer></script>
-  </head>
-  <body>
-    <div>
-      <h2>Events in JavaScript</h2>
-      <button>Click me</button>
-      <button>Click me too!</button>
-    </div>
-    <form action="">
-      <label for="title">Title</label>
-      <input type="text" id="title">
-      <button type="submit">Submit</button>
-    </form>
-    <ul>
-      <li>
-        <h2>Item 1</h2>
-        <p>Item description</p>
-      </li>
-      <li>
-        <h2>Item 2</h2>
-        <p>Item description</p>
-      </li>
-      <li>
-        <h2>Item 3</h2>
-        <p>Item description</p>
-      </li>
-      <li>
-        <h2>Item 4</h2>
-        <p>Item description</p>
-      </li>
-    </ul>
-    <section id="long-one">
-
-    </section>
-  </body>
-</html>
