+++ ProjectPlanner/index.html
@@ -6,7 +6,12 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge" />
     <title>Project Board</title>
     <link rel="stylesheet" href="assets/styles/app.css" />
-    <script src="assets/scripts/main.js" defer></script>
+    <!-- <script src="assets/scripts/utility/dom-helper.js" defer></script>
+    <script src="assets/scripts/app/component.js" defer></script>
+    <script src="assets/scripts/app/tooltip.js" defer type="module"></script>
+    <script src="assets/scripts/app/project-item.js" defer></script>
+    <script src="assets/scripts/app/project-list.js" defer></script> -->
+    <script src="assets/scripts/main.js" defer type="module"></script>
   </head>
   <body>
     <template  id="tooltip">
