--- ProjectPlanner/sizes.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Sizes & Positions</title>
-    <link rel="stylesheet" href="assets/styles/sizes.css" />
-    <script src="assets/scripts/sizes.js"></script>
-  </head>
-  <body>
-    <div id="main-box">
-      <h2>SIZES & POSITIONS</h2>
-      <p>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Nihil ad ea
-        tempora, repellat illum, autem culpa quidem dolor nulla praesentium
-        animi optio adipisci dicta error sint mollitia cum quos modi!
-      </p>
-      <p>
-        Lorem ipsum dolor sit amet consectetur adipisicing elit. Nihil ad ea
-        tempora, repellat illum, autem culpa quidem dolor nulla praesentium
-        animi optio adipisci dicta error sint mollitia cum quos modi!
-      </p>
-    </div>
-  </body>
-</html>
